Update app-service-web-configure-tls-mutual-auth.md
It's already possible to enable Client Certificates through the portal, so changing the text to reflect this:
From:
## Configure Web App for Client Certificate Authentication
To setup your web app to require client certificates you need to add the clientCertEnabled site setting for your web app and set it to true. This setting is not currently available through the management experience in the Portal, and the REST API will need to be used to accomplish this.
To:
## Configure Web App for Client Certificate Authentication
To setup your web app to require client certificates you need to add the clientCertEnabled site setting for your web app and set it to true. This setting is available through the management experience in the Portal, by going to SSL Settings and under Client Certificates setting "Incoming client certificates" to On.
Another option to accomplish this is to use the REST API.